Don't enable "-Werror" per default
Things are going to change. Warnings will be added, new compiler may become stricter
and considers a particular piece of code as something to warn about is not going to
change the quality of the software per-se; and while it’s true that fixing the
warnings early can save from failing further down the road, users often enough just
need the thing to build and work when they need it, and would prefer not to have to
fix a few warnings first.

So using "-Werror" per default on released code is not a good default.

This commit will make "-Werror" optional and disabled per default. Developers can
still pass "--enable-werror" option to configure to build with this flag during
development.